refactor: `SaveRecordingConfigPopup` component
