+++ app/src/views/features/sessions/SessionViewer/SaveRecordingConfigPopup.tsx
@@ -1,37 +1,179 @@
-import React from "react";
-import { useLocation } from "react-router-dom";
+import React, { useCallback, useEffect, useMemo, useState } from "react";
+import { useLocation, useNavigate, useParams } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
 import { CloseOutlined } from "@ant-design/icons";
-import { Button, Checkbox, Radio, RadioChangeEvent, Row } from "antd";
-import { SessionSaveMode } from "./types";
+import { Button, Checkbox, Radio, Row } from "antd";
+import { DebugInfo, SessionSaveMode } from "./types";
 import { CheckboxValueType } from "antd/lib/checkbox/Group";
+import {
+  compressEvents,
+  downloadSession,
+  getRecordingOptionsToSave,
+  getSessionEventsToSave,
+  getSessionRecordingOptions,
+  prepareSessionToExport,
+} from "./sessionEventsUtils";
+import { getSessionRecording, getSessionRecordingEvents } from "store/features/session-recording/selectors";
+import { toast } from "utils/Toast";
+import { getUserAuthDetails } from "store/selectors";
+import { getCurrentlyActiveWorkspace } from "store/features/teams/selectors";
+import { actions } from "store";
+import APP_CONSTANTS from "config/constants";
+import { AUTH } from "modules/analytics/events/common/constants";
+import { saveRecording } from "backend/sessionRecording/saveRecording";
+import PATHS from "config/constants/sub/paths";
+import {
+  trackDraftSessionSaveFailed,
+  trackDraftSessionSaved,
+} from "modules/analytics/events/features/sessionRecording";
 
 interface Props {
-  isSaving?: boolean;
-  sessionSaveMode: SessionSaveMode;
-  includedDebugInfo: CheckboxValueType[];
-  setIncludedDebugInfo: (mode: CheckboxValueType[]) => void;
-  onClose: (e: React.MouseEvent) => void;
-  isIncludeNetworkLogsDisabled: boolean;
-  isIncludeConsoleLogsDisabled: boolean;
-  handleSessionSaveModeChange?: (e: RadioChangeEvent) => void;
-  saveDraftSession?: (e: React.MouseEvent) => void;
-  handleDownloadFileClick: (e: React.MouseEvent) => void;
+  onClose: (e?: React.MouseEvent) => void;
 }
 
-const SaveRecordingConfigPopup: React.FC<Props> = ({
-  onClose,
-  isSaving = false,
-  sessionSaveMode,
-  saveDraftSession,
-  handleDownloadFileClick,
-  handleSessionSaveModeChange,
-  includedDebugInfo,
-  setIncludedDebugInfo,
-  isIncludeConsoleLogsDisabled,
-  isIncludeNetworkLogsDisabled,
-}) => {
+const { ACTION_LABELS: AUTH_ACTION_LABELS } = APP_CONSTANTS.AUTH;
+const defaultDebugInfo: CheckboxValueType[] = [DebugInfo.INCLUDE_NETWORK_LOGS, DebugInfo.INCLUDE_CONSOLE_LOGS];
+
+const SaveRecordingConfigPopup: React.FC<Props> = ({ onClose }) => {
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+  const { tabId } = useParams();
   const { pathname } = useLocation();
+  const user = useSelector(getUserAuthDetails);
+  const workspace = useSelector(getCurrentlyActiveWorkspace);
+  const sessionRecording = useSelector(getSessionRecording);
+  const sessionEvents = useSelector(getSessionRecordingEvents);
+  const [isSaving, setIsSaving] = useState(false);
+  const [sessionSaveMode, setSessionSaveMode] = useState<SessionSaveMode>(SessionSaveMode.ONLINE);
+  const [includedDebugInfo, setIncludedDebugInfo] = useState<CheckboxValueType[]>(defaultDebugInfo);
   const isDraftSession = pathname.includes("draft");
+  const isSessionLogOptionsAlreadySaved = tabId === "imported" || !isDraftSession;
+
+  const savedSessionRecordingOptions = useMemo(() => getSessionRecordingOptions(sessionRecording?.options), [
+    sessionRecording?.options,
+  ]);
+
+  const isIncludeNetworkLogsDisabled =
+    isSessionLogOptionsAlreadySaved && !savedSessionRecordingOptions.includes(DebugInfo.INCLUDE_NETWORK_LOGS);
+  const isIncludeConsoleLogsDisabled =
+    isSessionLogOptionsAlreadySaved && !savedSessionRecordingOptions.includes(DebugInfo.INCLUDE_CONSOLE_LOGS);
+
+  useEffect(() => {
+    if (isSessionLogOptionsAlreadySaved) {
+      setIncludedDebugInfo(savedSessionRecordingOptions);
+    }
+  }, [isSessionLogOptionsAlreadySaved, savedSessionRecordingOptions]);
+
+  const saveDraftSession = useCallback(
+    (e: React.MouseEvent) => {
+      if (!user?.loggedIn) {
+        dispatch(
+          actions.toggleActiveModal({
+            modalName: "authModal",
+            newValue: true,
+            newProps: {
+              authMode: AUTH_ACTION_LABELS.SIGN_UP,
+              src: window.location.href,
+              eventSource: AUTH.SOURCE.SAVE_DRAFT_SESSION,
+            },
+          })
+        );
+        return;
+      }
+
+      if (isSaving) {
+        return;
+      }
+
+      if (!sessionRecording?.name) {
+        toast.error("Name is required to save the recording.");
+        return;
+      }
+
+      const recordingOptionsToSave = getRecordingOptionsToSave(includedDebugInfo);
+
+      setIsSaving(true);
+      saveRecording(
+        user?.details?.profile?.uid,
+        workspace?.id,
+        sessionRecording,
+        compressEvents(getSessionEventsToSave(sessionEvents, recordingOptionsToSave)),
+        recordingOptionsToSave
+      ).then((response) => {
+        if (response?.success) {
+          onClose();
+          toast.success("Recording saved successfully");
+          trackDraftSessionSaved(
+            sessionRecording?.sessionAttributes?.duration,
+            recordingOptionsToSave,
+            SessionSaveMode.ONLINE
+          );
+          navigate(PATHS.SESSIONS.RELATIVE + "/saved/" + response?.firestoreId, {
+            replace: true,
+            state: { fromApp: true, viewAfterSave: true },
+          });
+        } else {
+          toast.error(response?.message);
+          trackDraftSessionSaveFailed(response?.message);
+          setIsSaving(false);
+        }
+      });
+    },
+    [
+      user?.loggedIn,
+      user?.details?.profile?.uid,
+      workspace?.id,
+      isSaving,
+      includedDebugInfo,
+      sessionEvents,
+      sessionRecording,
+      dispatch,
+      navigate,
+      onClose,
+    ]
+  );
+
+  const handleDownloadFileClick = useCallback(
+    (e: React.MouseEvent) => {
+      if (!user?.loggedIn) {
+        dispatch(
+          actions.toggleActiveModal({
+            modalName: "authModal",
+            newValue: true,
+            newProps: {
+              authMode: AUTH_ACTION_LABELS.SIGN_UP,
+              src: window.location.href,
+              eventSource: AUTH.SOURCE.SAVE_DRAFT_SESSION,
+            },
+          })
+        );
+        return;
+      }
+
+      setIsSaving(true);
+      const recordingOptionsToSave = getRecordingOptionsToSave(includedDebugInfo);
+      const events = compressEvents(getSessionEventsToSave(sessionEvents, recordingOptionsToSave));
+      const recording = {
+        name: sessionRecording.name,
+        options: { ...recordingOptionsToSave },
+        sessionAttributes: { ...sessionRecording.sessionAttributes },
+      };
+
+      prepareSessionToExport(events, recording)
+        .then((fileContent) => downloadSession(fileContent, sessionRecording.name))
+        .finally(() => {
+          toast.success("Recording downloaded successfully.");
+          onClose();
+          setIsSaving(false);
+          trackDraftSessionSaved(
+            sessionRecording.sessionAttributes?.duration,
+            recordingOptionsToSave,
+            SessionSaveMode.LOCAL
+          );
+        });
+    },
+    [dispatch, user?.loggedIn, sessionEvents, sessionRecording, includedDebugInfo, onClose]
+  );
 
   return (
     <div className="save-recording-config-popup">
@@ -45,7 +187,11 @@ const SaveRecordingConfigPopup: React.FC<Props> = ({
         {isDraftSession && (
           <div>
             <div className="config-label">Mode of saving the session</div>
-            <Radio.Group value={sessionSaveMode} onChange={handleSessionSaveModeChange} className="mode-radio-group">
+            <Radio.Group
+              value={sessionSaveMode}
+              className="mode-radio-group"
+              onChange={(e) => setSessionSaveMode(e.target.value)}
+            >
               <Radio value={SessionSaveMode.ONLINE}>Save online</Radio>
               <Radio value={SessionSaveMode.LOCAL}>Download as a file</Radio>
             </Radio.Group>
@@ -77,9 +223,11 @@ const SaveRecordingConfigPopup: React.FC<Props> = ({
           type="primary"
           className="ml-auto"
           loading={isSaving}
-          onClick={sessionSaveMode === SessionSaveMode.ONLINE ? saveDraftSession : handleDownloadFileClick}
+          onClick={
+            isDraftSession && sessionSaveMode === SessionSaveMode.ONLINE ? saveDraftSession : handleDownloadFileClick
+          }
         >
-          {sessionSaveMode === SessionSaveMode.ONLINE ? "Save" : "Download"}
+          {isDraftSession && sessionSaveMode === SessionSaveMode.ONLINE ? "Save" : "Download"}
         </Button>
       </Row>
     </div>
